requirements.txt
fastapi
pillow
numpy
opencv-python-headless